moved ChatMessage + related structs into chat_message.rs; renames within chat_message;
