+++ src/call_validation.rs
@@ -1,15 +1,12 @@
-use serde::{Deserialize, Deserializer, Serialize};
+use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::hash::Hash;
-use std::sync::{Arc, RwLock, RwLockWriteGuard};
 use axum::http::StatusCode;
 use indexmap::IndexMap;
-use regex::Regex;
 use ropey::Rope;
-use tokenizers::Tokenizer;
 
 use crate::custom_error::ScratchError;
-use crate::scratchpads::chat_utils_limit_history::calculate_image_tokens_openai;
+use crate::scratchpads::chat_message::ChatMessage;
 
 
 #[derive(Debug, Serialize, Deserialize, Clone, Default)]
@@ -99,241 +96,6 @@ pub enum ContextEnum {
 
 fn default_gradient_type_value() -> i32 {
     -1
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
-pub enum ChatMultimodalElement {
-    MultimodalTextElement(MultimodalTextElement),
-    MultiModalImageURLElement(MultimodalImageURLElement),
-}
-
-impl TryFrom<serde_json::Value> for ChatMultimodalElement {
-    type Error = String;
-
-    fn try_from(value: serde_json::Value) -> Result<Self, Self::Error> {
-        let content_type = value.get("type").ok_or("field 'type' is missing".to_string())?;
-        let content_type_str = content_type.as_str().ok_or("field 'type' is not a string".to_string())?;
-        
-        match content_type_str {
-            "text" => Ok(ChatMultimodalElement::MultimodalTextElement(serde_json::from_value(value).map_err(|e| e.to_string())?)),
-            "image_url" => Ok(ChatMultimodalElement::MultiModalImageURLElement(serde_json::from_value(value).map_err(|e| e.to_string())?)),
-            _ => Err(format!("invalid type: {}; must be one of: text, image_url", content_type_str)),
-        }
-    }
-}
-
-impl Default for ChatMultimodalElement {
-    fn default() -> Self {
-        ChatMultimodalElement::MultimodalTextElement(MultimodalTextElement {
-            content_type: "text".to_string(),
-            text: String::new(),
-        })
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Default)]
-pub struct MultimodalTextElement {
-    #[serde(rename = "type")]
-    pub content_type: String,
-    pub text: String,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Default)]
-pub struct MultimodalImageURLElement {
-    #[serde(rename = "type")]
-    pub content_type: String,
-    pub image_url: MultimodalImageURLElementImageURL,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Default)]
-pub struct MultimodalImageURLElementImageURL {
-    pub url: String,
-    #[serde(default = "default_detail")]
-    pub detail: String,
-}
-
-fn default_detail() -> String {
-    "high".to_string()
-}
-
-// todo: images via links are yet not implemented: unclear how to calculate tokens
-fn parse_image_b64_from_image_url(image_url: &str) -> Option<String> {
-    let re = Regex::new(r"data:image/(png|jpeg|jpg|webp|gif);base64,([A-Za-z0-9+/=]+)").unwrap();
-    re.captures(image_url).and_then(|captures| {
-        captures.get(2).map(|m| m.as_str().to_string())
-    })
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
-#[serde(untagged)]
-pub enum ChatContent {
-    SimpleText(String),
-    Multimodal(Vec<ChatMultimodalElement>),
-}
-
-impl Default for ChatContent {
-    fn default() -> Self {
-        ChatContent::SimpleText(String::new())
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct ChatToolFunction {
-    pub arguments: String,
-    pub name: String,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct ChatToolCall {
-    pub id: String,
-    pub function: ChatToolFunction,
-    #[serde(rename = "type")]
-    pub tool_type: String,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone, Default)]
-pub struct ChatUsage {
-    pub prompt_tokens: usize,
-    pub completion_tokens: usize,
-    pub total_tokens: usize,   // TODO: remove (can produce self-contradictory data when prompt+completion != total)
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone, Default)]
-pub struct ChatMessage {
-    pub role: String,
-    #[serde(default, deserialize_with="deserialize_chat_content")]
-    pub content: ChatContent,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub tool_calls: Option<Vec<ChatToolCall>>,
-    #[serde(default)]
-    pub tool_call_id: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub usage: Option<ChatUsage>,
-}
-
-pub fn chat_content_from_value(value: serde_json::Value) -> Result<ChatContent, String> {
-    fn validate_multimodal_element(element: &ChatMultimodalElement) -> Result<(), String> {
-        match element {
-            ChatMultimodalElement::MultimodalTextElement(el) => {
-                if el.content_type != "text" {
-                    return Err("Invalid multimodal element: type must be `text`".to_string());
-                }
-            },
-            ChatMultimodalElement::MultiModalImageURLElement(el) => {
-                if el.content_type != "image_url" {
-                    return Err("Invalid multimodal element: type must be `image_url`".to_string());
-                }
-                if parse_image_b64_from_image_url(&el.image_url.url).is_none() {
-                    return Err("Invalid image URL in MultimodalImageURLElement: must pass regexp `data:image/(png|jpeg|jpg|webp|gif);base64,([A-Za-z0-9+/=]+)`".to_string());
-                }
-            }
-        };
-        Ok(())
-    }
-    
-    match value {
-        serde_json::Value::String(s) => Ok(ChatContent::SimpleText(s)),
-        serde_json::Value::Array(array) => {
-            let mut elements = vec![];
-            for (idx, item) in array.iter().enumerate() {
-                let element: ChatMultimodalElement = ChatMultimodalElement::try_from(item.clone())
-                    .map_err(|e| format!("Error deserializing element at index {}: {}", idx, e))?;
-                validate_multimodal_element(&element)
-                    .map_err(|e| format!("Validation error for element at index {}: {}", idx, e))?;
-                elements.push(element);
-            }
-            if elements.len() == 1 {
-                if let ChatMultimodalElement::MultimodalTextElement(el) = &elements[0] {
-                    return Ok(ChatContent::SimpleText(el.text.clone()));
-                }
-            }
-            Ok(ChatContent::Multimodal(elements))
-        },
-        _ => Err("deserialize_chat_content() can't parse content".to_string()),
-    }
-}
-
-fn deserialize_chat_content<'de, D>(deserializer: D) -> Result<ChatContent, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let value: serde_json::Value = Deserialize::deserialize(deserializer)?;
-    chat_content_from_value(value).map_err(serde::de::Error::custom)
-}
-
-impl ChatMessage {
-    pub fn new(role: String, content: String) -> Self {
-        ChatMessage { 
-            role, 
-            content: ChatContent::SimpleText(content),
-            ..Default::default()
-        }
-    }
-
-    pub fn drop_usage(&self) -> ChatMessage {
-        ChatMessage {
-            role: self.role.clone(),
-            content: self.content.clone(),
-            tool_calls: self.tool_calls.clone(),
-            tool_call_id: self.tool_call_id.clone(),
-            usage: None,
-        }
-    }
-}
-
-impl ChatContent {
-    pub fn content_text_only(&self) -> String {
-        match self {
-            ChatContent::SimpleText(text) => text.clone(),
-            ChatContent::Multimodal(elements) => {
-                elements
-                    .iter()
-                    .filter_map(|element| {
-                        match element {
-                            ChatMultimodalElement::MultimodalTextElement(el) => Some(el.text.clone()),
-                            _ => None,
-                        }
-                    })
-                    .collect::<Vec<String>>()
-                    .join("\n\n")
-            }
-        }
-    }
-
-    pub fn size_estimate(&self) -> usize {
-        match self {
-            ChatContent::SimpleText(text) => text.len(),
-            _ => unreachable!()
-        }
-    }
-    
-    pub fn count_tokens(&self, tokenizer: Arc<RwLock<Tokenizer>>) -> Result<i32, String> {
-        fn count_tokens_simple_text(tokenizer_lock: &RwLockWriteGuard<Tokenizer>, text: &str) -> Result<i32, String> {
-            tokenizer_lock.encode(text, false)
-                .map(|tokens|tokens.len() as i32)
-               .map_err(|e|format!("Tokenizing error: {e}"))
-        }
-        let tokenizer_lock = tokenizer.write().unwrap();
-        match self {
-            ChatContent::SimpleText(text) => count_tokens_simple_text(&tokenizer_lock, text),
-            ChatContent::Multimodal(elements) => {
-                let mut tcnt = 0;
-                for e in elements {
-                    tcnt += match e {
-                        ChatMultimodalElement::MultimodalTextElement(el) => count_tokens_simple_text(&tokenizer_lock, el.text.as_str())?,
-                        ChatMultimodalElement::MultiModalImageURLElement(el) => {
-                            if let Some(image_base64) = parse_image_b64_from_image_url(el.image_url.url.as_str()) {
-                                calculate_image_tokens_openai(&image_base64, &el.image_url.detail)?
-                            } else {
-                                0
-                            }
-                        }
-                    };
-                }
-                Ok(tcnt)
-            }
-        }
-    }
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
@@ -437,19 +199,6 @@ mod tests {
     use std::collections::HashMap;
     use crate::call_validation::{CodeCompletionInputs, CursorPosition, SamplingParameters};
     use super::*;
-
-    #[test]
-    fn test_parse_image_b64_from_image_url() {
-        let image_url = "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAUA";
-        let expected_base64 = "iVBORw0KGgoAAAANSUhEUgAAAAUA";
-        assert_eq!(parse_image_b64_from_image_url(image_url), Some(expected_base64.to_string()));
-
-        let invalid_image_url = "data:image/png;base64,";
-        assert_eq!(parse_image_b64_from_image_url(invalid_image_url), None);
-
-        let non_matching_url = "https://example.com/image.png";
-        assert_eq!(parse_image_b64_from_image_url(non_matching_url), None);
-    }
 
     #[test]
     fn test_valid_post1() {
