+++ src/tools/tool_ast_reference.rs
@@ -7,7 +7,8 @@ use tokio::sync::Mutex as AMutex;
 
 use crate::at_commands::at_commands::AtCommandsContext;
 use crate::tools::tools_description::Tool;
-use crate::call_validation::{ChatMessage, ChatContent, ContextEnum, ContextFile};
+use crate::call_validation::{ContextEnum, ContextFile};
+use crate::scratchpads::chat_message::{ChatContent, ChatMessage};
 use crate::tools::tool_ast_definition::there_are_definitions_with_similar_names_though;
