+++ src/tools/tool_knowledge.rs
@@ -8,7 +8,8 @@ use async_trait::async_trait;
 
 use crate::at_commands::at_commands::AtCommandsContext;
 use crate::tools::tools_description::Tool;
-use crate::call_validation::{ChatMessage, ChatContent, ContextEnum};
+use crate::call_validation::ContextEnum;
+use crate::scratchpads::chat_message::{ChatContent, ChatMessage};
 use crate::vecdb::vdb_highlev::memories_search;
 // use crate::vecdb::vdb_highlev::ongoing_find;
