+++ src/tools/tool_patch_aux/model_based_edit/model_execution.rs
@@ -1,13 +1,3 @@
-use crate::at_commands::at_commands::AtCommandsContext;
-use crate::cached_tokenizers::cached_tokenizer;
-use crate::call_validation::{ChatMessage, ChatUsage, DiffChunk};
-use crate::global_context::{try_load_caps_quickly_if_not_present, GlobalContext};
-use crate::scratchpads::scratchpad_utils::count_tokens;
-use crate::subchat::subchat_single;
-use crate::tools::tool_patch_aux::fs_utils::read_file;
-use crate::tools::tool_patch_aux::model_based_edit::blocks_of_code_parser::BlocksOfCodeParser;
-use crate::tools::tool_patch_aux::model_based_edit::whole_file_parser::WholeFileParser;
-use crate::tools::tool_patch_aux::tickets_parsing::TicketToApply;
 use std::path::PathBuf;
 use std::sync::Arc;
 use std::sync::RwLock as StdRwLock;
@@ -15,6 +5,18 @@ use tokenizers::Tokenizer;
 use tokio::sync::Mutex as AMutex;
 use tokio::sync::RwLock as ARwLock;
 use tracing::{info, warn};
+
+use crate::at_commands::at_commands::AtCommandsContext;
+use crate::cached_tokenizers::cached_tokenizer;
+use crate::call_validation::DiffChunk;
+use crate::global_context::{try_load_caps_quickly_if_not_present, GlobalContext};
+use crate::scratchpads::scratchpad_utils::count_tokens;
+use crate::subchat::subchat_single;
+use crate::tools::tool_patch_aux::fs_utils::read_file;
+use crate::tools::tool_patch_aux::model_based_edit::blocks_of_code_parser::BlocksOfCodeParser;
+use crate::tools::tool_patch_aux::model_based_edit::whole_file_parser::WholeFileParser;
+use crate::tools::tool_patch_aux::tickets_parsing::TicketToApply;
+use crate::scratchpads::chat_message::{ChatMessage, ChatUsage};
 
 
 const DEBUG: bool = true;
