+++ src/tools/tool_web.rs
@@ -7,7 +7,8 @@ use tokio::sync::Mutex as AMutex;
 use crate::at_commands::at_commands::AtCommandsContext;
 use crate::at_commands::at_web::execute_at_web;
 use crate::tools::tools_description::Tool;
-use crate::call_validation::{ChatMessage, ChatContent, ContextEnum};
+use crate::call_validation::ContextEnum;
+use crate::scratchpads::chat_message::{ChatContent, ChatMessage};
 
 
 pub struct ToolWeb;
